style: move InLevelUI.cs to UI folder
