+++ Assets/Scripts/Camera/CameraAspectRatioClamper.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 52656acffe9f5374a94d759ccbdaa636
+guid: 076b98d1f7588934ea89a5e06e51e86c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
