--- Assets/Scripts/UIMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6298b102c3e43d647a64c14081f0eab5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
